+++ packages/client/lib/rpc/modules/eth.ts
@@ -12,7 +12,7 @@ import {
   Account,
   Address,
   bufferToHex,
-  bnToHex,
+  bigIntToHex,
   intToHex,
   rlp,
   toBuffer,
@@ -158,7 +158,7 @@ const jsonRpcBlock = async (
     receiptsRoot: header.receiptTrie!,
     miner: header.coinbase!,
     difficulty: header.difficulty!,
-    totalDifficulty: bnToHex(td),
+    totalDifficulty: bigIntToHex(td),
     extraData: header.extraData!,
     size: intToHex(Buffer.byteLength(JSON.stringify(json))),
     gasLimit: header.gasLimit!,
@@ -177,7 +177,7 @@ const jsonRpcTx = (tx: TypedTransaction, block?: Block, txIndex?: number): JsonR
   const txJSON = tx.toJSON()
   return {
     blockHash: block ? bufferToHex(block.hash()) : null,
-    blockNumber: block ? bnToHex(block.header.number) : null,
+    blockNumber: block ? bigIntToHex(block.header.number) : null,
     from: tx.getSenderAddress().toString(),
     gas: txJSON.gasLimit!,
     gasPrice: txJSON.gasPrice ?? txJSON.maxFeePerGas!,
@@ -213,7 +213,7 @@ const jsonRpcLog = async (
   transactionIndex: txIndex !== undefined ? intToHex(txIndex) : null,
   transactionHash: tx ? bufferToHex(tx.hash()) : null,
   blockHash: block ? bufferToHex(block.hash()) : null,
-  blockNumber: block ? bnToHex(block.header.number) : null,
+  blockNumber: block ? bigIntToHex(block.header.number) : null,
   address: bufferToHex(log[0]),
   topics: log[1].map((t) => bufferToHex(t as Buffer)),
   data: bufferToHex(log[2]),
@@ -235,12 +235,12 @@ const jsonRpcReceipt = async (
   transactionHash: bufferToHex(tx.hash()),
   transactionIndex: intToHex(txIndex),
   blockHash: bufferToHex(block.hash()),
-  blockNumber: bnToHex(block.header.number),
+  blockNumber: bigIntToHex(block.header.number),
   from: tx.getSenderAddress().toString(),
   to: tx.to?.toString() ?? null,
   cumulativeGasUsed: bufferToHex(receipt.gasUsed),
-  effectiveGasPrice: bnToHex(effectiveGasPrice),
-  gasUsed: bnToHex(gasUsed),
+  effectiveGasPrice: bigIntToHex(effectiveGasPrice),
+  gasUsed: bigIntToHex(gasUsed),
   contractAddress: contractAddress?.toString() ?? null,
   logs: await Promise.all(
     receipt.logs.map((l, i) => jsonRpcLog(l, block, tx, txIndex, logIndex + i))
@@ -418,7 +418,7 @@ export class Eth {
    * @param params An empty array
    */
   async blockNumber(_params = []) {
-    return bnToHex(this._chain.headers.latest?.number ?? BigInt(0))
+    return bigIntToHex(this._chain.headers.latest?.number ?? BigInt(0))
   }
 
   /**
@@ -471,7 +471,7 @@ export class Eth {
    */
   async chainId(_params = []) {
     const chainId = this._chain.config.chainCommon.chainId()
-    return bnToHex(chainId)
+    return bigIntToHex(chainId)
   }
 
   /**
@@ -543,7 +543,7 @@ export class Eth {
     const vm = this._vm.copy()
     await vm.stateManager.setStateRoot(block.header.stateRoot)
     const account = await vm.stateManager.getAccount(address)
-    return bnToHex(account.balance)
+    return bigIntToHex(account.balance)
   }
 
   /**
@@ -685,7 +685,7 @@ export class Eth {
 
     const address = Address.fromString(addressHex)
     const account: Account = await vm.stateManager.getAccount(address)
-    return bnToHex(account.nonce)
+    return bigIntToHex(account.nonce)
   }
 
   /**
@@ -967,14 +967,14 @@ export class Eth {
     }
 
     const currentBlockHeader = this._chain.headers?.latest ?? (await this._chain.getLatestHeader())
-    const currentBlock = bnToHex(currentBlockHeader.number)
+    const currentBlock = bigIntToHex(currentBlockHeader.number)
 
     const synchronizer = this.client.services[0].synchronizer
-    const startingBlock = bnToHex(synchronizer.startingBlock)
+    const startingBlock = bigIntToHex(synchronizer.startingBlock)
 
     let highestBlock
     if (synchronizer.syncTargetHeight) {
-      highestBlock = bnToHex(synchronizer.syncTargetHeight)
+      highestBlock = bigIntToHex(synchronizer.syncTargetHeight)
     } else {
       const bestPeer = synchronizer.best()
       if (!bestPeer) {
@@ -990,7 +990,7 @@ export class Eth {
           message: `highest block header unavailable`,
         }
       }
-      highestBlock = bnToHex(highestBlockHeader.number)
+      highestBlock = bigIntToHex(highestBlockHeader.number)
     }
 
     return { startingBlock, currentBlock, highestBlock }